Feat: ArgumentResolver 적용 및 테스트 수정

+++ src/main/java/gift/service/WishService.java
@@ -1,10 +1,9 @@
 package gift.service;
 
-import gift.domain.Member;
+import gift.config.UnAuthorizationException;
 import gift.domain.Product;
 import gift.domain.Wish;
 import gift.dto.*;
-import gift.repository.MemberRepository;
 import gift.repository.ProductRepository;
 import gift.repository.WishRepository;
 import org.springframework.stereotype.Service;
@@ -18,52 +17,50 @@ public class WishService {
 
     private final ProductRepository productRepository;
     private final WishRepository wishRepository;
-    private final MemberRepository memberRepository;
 
-    public WishService(ProductRepository productRepository, WishRepository wishRestController, MemberRepository memberRepository) {
+    public WishService(ProductRepository productRepository, WishRepository wishRestController) {
         this.productRepository = productRepository;
         this.wishRepository = wishRestController;
-        this.memberRepository = memberRepository;
     }
 
     public List<Product> productList() {
         return productRepository.findAll();
     }
 
-    public CreateWishResponse addWishProduct(CreateWishRequest request) {
-        Wish wish = wishRepository.save(request);
+    public CreateWishResponse addWishProduct(CreateWishRequest request, Long loginMemberId) {
+        Wish wish = wishRepository.save(request.productId(), loginMemberId, request.quantity());
         return new CreateWishResponse(wish.getId(), wish.getMemberId(), wish.getProductId(), wish.getQuantity());
     }
 
     public List<WishResponse> getMeberWishList(Long memberId) {
-        findMemberOrThrow(memberId);
         return wishRepository.findAllByMember(memberId);
     }
 
-    public void delete(Long wishId) {
-        findByIdOrThrow(wishId);
+    public void delete(Long wishId, Long memberId) {
+        Wish wishProduct = findByIdOrThrow(wishId);
+        checkAuthorization(wishProduct, memberId, "삭제 권한 없음");
         wishRepository.delete(wishId);
     }
 
-    public UpdateWishResponse updateQuantity(UpdateWishRequest request, Long wishId) {
+    public UpdateWishResponse updateQuantity(UpdateWishRequest request, Long wishId, Long memberId) {
         Wish wishProduct = findByIdOrThrow(wishId);
-        wishRepository.update(request, wishId);
+        checkAuthorization(wishProduct, memberId, "수정 권한 없음");
+
+        wishRepository.update(request.quantity(), wishId);
         return new UpdateWishResponse(wishProduct.getId(), wishProduct.getProductId(), request.quantity());
+    }
+
+    private static void checkAuthorization(Wish wishProduct, Long memberId, String exMessage) {
+        if (!wishProduct.getMemberId().equals(memberId)) {
+            throw new UnAuthorizationException(exMessage);
+        }
     }
 
     private Wish findByIdOrThrow(Long wishId) {
         Optional<Wish> findWishProduct = wishRepository.findById(wishId);
         if (findWishProduct.isEmpty()) {
-            throw new NoSuchElementException("존재하지 않는 위시리스트 상품입니다.");
+            throw new NoSuchElementException("존재하지 않는 위시리스트 상품");
         }
         return findWishProduct.get();
-    }
-
-    private Member findMemberOrThrow(Long memberId) {
-        Optional<Member> findMember = memberRepository.findById(memberId);
-        if (findMember.isEmpty()) {
-            throw new NoSuchElementException("존재하지 않는 멤버입니다.");
-        }
-        return findMember.get();
     }
 }
